fix: token config during setup

+++ solidity/contracts/interfaces/IMoneyMarketReader.sol
@@ -7,8 +7,8 @@ interface IMoneyMarketReader {
     uint256 ibTotalSupply;
     uint256 ibTotalAsset;
     address ibAddress;
-    uint256 ibCollateralFactor;
-    uint256 ibBorrowingFactor;
+    uint16 ibCollateralFactor;
+    uint16 ibBorrowingFactor;
     // ---- Token Config ---- //
     uint8 tierAsUInt;
     uint16 collateralFactor;
